⚡separated schema definitions for video embed components for quicker hot reloads
